+++ packages/atproto/test/src/servers/servers_service_test.dart
@@ -16,51 +16,9 @@ import 'package:atproto/src/entities/created_invite_codes.dart';
 import 'package:atproto/src/entities/current_session.dart';
 import 'package:atproto/src/entities/invite_codes.dart';
 import 'package:atproto/src/entities/server_info.dart';
-import 'package:atproto/src/entities/session.dart';
 import 'package:atproto/src/servers/servers_service.dart';
 
 void main() {
-  group('.createSession', () {
-    test('normal case', () async {
-      final response = await createSession(
-        identifier: 'shinyakato.dev',
-        password: '1234',
-        mockedPostClient: atp_test.createMockedPostClient(
-          'test/src/servers/data/create_session.json',
-        ),
-      );
-
-      expect(response, isA<core.XRPCResponse>());
-      expect(response.data, isA<Session>());
-    });
-
-    test('when unauthorized', () async {
-      atp_test.expectUnauthorizedException(
-        () async => await createSession(
-          identifier: 'shinyakato.dev',
-          password: '1234',
-          mockedPostClient: atp_test.createMockedPostClient(
-            'test/src/data/error.json',
-            statusCode: 401,
-          ),
-        ),
-      );
-    });
-
-    test('when rate limit exceeded', () async {
-      atp_test.expectRateLimitExceededException(
-        () async => await createSession(
-          identifier: 'shinyakato.dev',
-          password: '1234',
-          mockedPostClient: atp_test.createMockedPostClient(
-            'test/src/data/error.json',
-            statusCode: 429,
-          ),
-        ),
-      );
-    });
-  });
-
   group('.findCurrentSession', () {
     test('normal case', () async {
       final servers = ServersService(
@@ -163,7 +121,7 @@ void main() {
       );
 
       expect(response, isA<core.XRPCResponse>());
-      expect(response.data, isA<Session>());
+      expect(response.data, isA<core.Session>());
     });
 
     test('when unauthorized', () async {