+++ packages/atproto_core/lib/src/const.dart
@@ -2,5 +2,17 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+// 📦 Package imports:
+import 'package:xrpc/xrpc.dart' as xrpc;
+
 /// The unmodifiable empty JSON.
 const emptyJson = <String, dynamic>{};
+
+/// The default HTTP protocol.
+const defaultProtocol = xrpc.Protocol.https;
+
+/// The default timeout duration.
+const defaultTimeout = Duration(seconds: 30);
+
+/// `bsky.social`
+const defaultService = 'bsky.social';